fix: resolve Float/Double type mismatches in PdfContentCalculator and PdfRenderer

+++ app/src/main/kotlin/com/sulat/ai/document/renderer/PdfContentCalculator.kt
@@ -46,7 +46,7 @@ class PdfContentCalculator(
             cursorY = page.marginTopPt
         }
 
-        fun addSpacer(spacerPt: Double = PdfTextRole.SPACER.style.fontSizePt.toDouble()) {
+        fun addSpacer(spacerPt: Double = PdfTextRole.SPACER.style.fontSizePt) {
             if (!checkSpace(spacerPt)) newPage()
             cursorY += spacerPt
         }
